Fix issues related to AgentTimeCharge
