+++ apps/studio.giselles.ai/app/(main)/settings/team/v2/agent-time-charge.tsx
@@ -3,7 +3,7 @@ import { fetchCurrentTeam, isProPlan } from "@/services/teams";
 import {
 	AgentTimeUsageForFreePlan,
 	AgentTimeUsageForProPlan,
-} from "./components/agent-time-usage";
+} from "../components/v2/agent-time-usage";
 
 export async function AgentTimeCharge() {
 	const currentTeam = await fetchCurrentTeam();
@@ -13,7 +13,7 @@ export async function AgentTimeCharge() {
 	const usedMinutes = Math.ceil((timeChargeMs / 1000 / 60) * 100) / 100;
 
 	return (
-		<div className="bg-transparent rounded-[16px] border border-black-70 py-[16px] px-[24px] w-full gap-[16px] grid">
+		<div className="bg-transparent rounded-[8px] border-[0.5px] border-black-400 px-[24px] pt-[16px] pb-[24px] w-full gap-[24px] grid">
 			{currentTeamIsPro ? (
 				<AgentTimeUsageForProPlan usedMinutes={usedMinutes} />
 			) : (
